Small fixes

+++ gogoreader/background.go
@@ -8,11 +8,11 @@ import (
 )
 
 func AverageColor(pictureData *pixel.PictureData, rect image.Rectangle) pixel.RGBA {
-	step := 3
+	step := 3.0
 	var count, sr, sg, sb float64
-	for x := rect.Min.X; x < rect.Max.X; x += step {
-		for y := rect.Min.Y; y < rect.Max.Y; y += step {
-			rgba := pictureData.Color(pixel.Vec{X: float64(x), Y: float64(y)})
+	for x := float64(rect.Min.X); x < float64(rect.Max.X); x += step {
+		for y := float64(rect.Min.Y); y < float64(rect.Max.Y); y += step {
+			rgba := pictureData.Color(pixel.Vec{X: x, Y: y})
 			r, g, b := rgba.R, rgba.G, rgba.B
 			if r > 0.95 && g > 0.95 && b > 0.95 {
 				// ignore white pixels
@@ -33,13 +33,13 @@ func AverageColor(pictureData *pixel.PictureData, rect image.Rectangle) pixel.RG
 }
 
 func ProminentColor(pictureData *pixel.PictureData, rect image.Rectangle) pixel.RGBA {
-	step := 3
+	step := 3.0
 
 	colorsCount := make(map[pixel.RGBA]uint)
 	currentMax := uint(0)
 	prominentColor := pixel.RGB(0, 0, 0)
-	for x := rect.Min.X; x < rect.Max.X; x += step {
-		for y := rect.Min.Y; y < rect.Max.Y; y += step {
+	for x := float64(rect.Min.X); x < float64(rect.Max.X); x += step {
+		for y := float64(rect.Min.Y); y < float64(rect.Max.Y); y += step {
 			rgba := pictureData.Color(pixel.Vec{X: float64(x), Y: float64(y)})
 			r, g, b := rgba.R, rgba.G, rgba.B
 			if r > 0.95 && g > 0.95 && b > 0.95 {
